homepage restructure and context
